+++ src/main/kotlin/no/nav/familie/ef/infotrygd/feed/Launcher.kt
@@ -1,10 +1,10 @@
 package no.nav.familie.ef.infotrygd.feed
 
-import no.nav.familie.ba.infotrygd.feed.config.ApplicationConfig
+import no.nav.familie.ef.infotrygd.feed.config.ApplicationConfig
 import org.springframework.boot.SpringApplication
 import org.springframework.boot.autoconfigure.SpringBootApplication
 
-@SpringBootApplication(scanBasePackages = ["no.nav.familie.ba.infotrygd.feed"])
+@SpringBootApplication(scanBasePackages = ["no.nav.familie.ef.infotrygd.feed"])
 class Launcher
 
 fun main(args: Array<String>) {
