+++ src/main/kotlin/no/nav/familie/ef/infotrygd/feed/database/FeedRepository.kt
@@ -2,7 +2,7 @@ package no.nav.familie.ef.infotrygd.feed.database
 
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.Query
-import no.nav.familie.ba.infotrygd.feed.rest.dto.Type
+import no.nav.familie.ef.infotrygd.feed.rest.dto.Type
 import org.springframework.data.domain.Pageable
 
 interface FeedRepository : JpaRepository<Feed, Long> {
