+++ src/main/kotlin/no/nav/familie/ef/infotrygd/feed/rest/OpprettFeedController.kt
@@ -1,9 +1,9 @@
 package no.nav.familie.ef.infotrygd.feed.rest
 
-import no.nav.familie.ba.infotrygd.feed.rest.dto.FødselsDto
-import no.nav.familie.ba.infotrygd.feed.rest.dto.Type
-import no.nav.familie.ba.infotrygd.feed.rest.dto.VedtakDto
-import no.nav.familie.ba.infotrygd.feed.service.InfotrygdFeedService
+import no.nav.familie.ef.infotrygd.feed.rest.dto.FødselsDto
+import no.nav.familie.ef.infotrygd.feed.rest.dto.Type
+import no.nav.familie.ef.infotrygd.feed.rest.dto.VedtakDto
+import no.nav.familie.ef.infotrygd.feed.service.InfotrygdFeedService
 import no.nav.familie.kontrakter.felles.Ressurs
 import no.nav.security.token.support.core.api.ProtectedWithClaims
 import org.slf4j.LoggerFactory
